delete old nav
