+++ web/packages/teleport/src/Navigation/CategoryIcon.tsx
