+++ web/packages/teleport/src/Navigation/ResourcesSection.tsx
