+++ web/packages/teleport/src/TopBar/index.ts
@@ -17,3 +17,4 @@
  */
 
 export { TopBar, TopBarContainer } from './TopBar';
+export type { TopBarProps } from './TopBar';
