+++ web/packages/teleport/src/features.tsx
@@ -40,10 +40,9 @@ import { IntegrationEnroll } from '@gravitational/teleport/src/Integrations/Enro
 import cfg, { Cfg } from 'teleport/config';
 import { IntegrationStatus } from 'teleport/Integrations/IntegrationStatus';
 import {
-  ManagementSection,
   NavigationCategory,
+  NavigationCategory as SideNavigationCategory,
 } from 'teleport/Navigation/categories';
-import { NavigationCategory as SideNavigationCategory } from 'teleport/Navigation/SideNavigation/categories';
 
 import { LockedAccessRequests } from './AccessRequests';
 import { AccountPage } from './Account';
@@ -103,9 +102,8 @@ class AccessRequests implements TeleportFeature {
 }
 
 export class FeatureJoinTokens implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Access;
-  sideNavCategory = SideNavigationCategory.Access;
+  category = NavigationCategory.Access;
+
   navigationItem = {
     title: NavTitle.JoinTokens,
     icon: Key,
@@ -172,8 +170,7 @@ export class FeatureUnifiedResources implements TeleportFeature {
 }
 
 export class FeatureSessions implements TeleportFeature {
-  category = NavigationCategory.Resources;
-  sideNavCategory = SideNavigationCategory.Audit;
+  category = NavigationCategory.Audit;
 
   route = {
     title: 'Active Sessions',
@@ -205,9 +202,7 @@ export class FeatureSessions implements TeleportFeature {
 // - Access
 
 export class FeatureUsers implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Access;
-  sideNavCategory = SideNavigationCategory.Access;
+  category = NavigationCategory.Access;
 
   route = {
     title: 'Manage Users',
@@ -238,9 +233,7 @@ export class FeatureUsers implements TeleportFeature {
 }
 
 export class FeatureBots implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Access;
-  sideNavCategory = SideNavigationCategory.Access;
+  category = NavigationCategory.Access;
 
   route = {
     title: 'Manage Bots',
@@ -274,9 +267,7 @@ export class FeatureBots implements TeleportFeature {
 }
 
 export class FeatureAddBots implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Access;
-  sideNavCategory = SideNavigationCategory.AddNew;
+  category = NavigationCategory.AddNew;
 
   route = {
     title: 'Bot',
@@ -305,9 +296,7 @@ export class FeatureAddBots implements TeleportFeature {
 }
 
 export class FeatureRoles implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Permissions;
-  sideNavCategory = SideNavigationCategory.Access;
+  category = NavigationCategory.Access;
 
   route = {
     title: 'Manage User Roles',
@@ -334,9 +323,7 @@ export class FeatureRoles implements TeleportFeature {
 }
 
 export class FeatureAuthConnectors implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Access;
-  sideNavCategory = SideNavigationCategory.Access;
+  category = NavigationCategory.Access;
 
   route = {
     title: 'Manage Auth Connectors',
@@ -365,9 +352,7 @@ export class FeatureAuthConnectors implements TeleportFeature {
 }
 
 export class FeatureLocks implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Identity;
-  sideNavCategory = SideNavigationCategory.Identity;
+  category = NavigationCategory.Identity;
 
   route = {
     title: 'Session & Identity Locks',
@@ -411,9 +396,7 @@ export class FeatureNewLock implements TeleportFeature {
 }
 
 export class FeatureDiscover implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Access;
-  sideNavCategory = SideNavigationCategory.AddNew;
+  category = NavigationCategory.AddNew;
   standalone = true;
 
   route = {
@@ -443,9 +426,7 @@ export class FeatureDiscover implements TeleportFeature {
 }
 
 export class FeatureIntegrations implements TeleportFeature {
-  sideNavCategory = SideNavigationCategory.Access;
-  category = NavigationCategory.Management;
-  section = ManagementSection.Access;
+  category = NavigationCategory.Access;
 
   hasAccess(flags: FeatureFlags) {
     // if feature hiding is enabled, only show
@@ -479,9 +460,7 @@ export class FeatureIntegrations implements TeleportFeature {
 }
 
 export class FeatureIntegrationEnroll implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Access;
-  sideNavCategory = SideNavigationCategory.AddNew;
+  category = NavigationCategory.AddNew;
 
   route = {
     title: 'Integration',
@@ -516,9 +495,7 @@ export class FeatureIntegrationEnroll implements TeleportFeature {
 // - Activity
 
 export class FeatureRecordings implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Activity;
-  sideNavCategory = SideNavigationCategory.Audit;
+  category = NavigationCategory.Audit;
 
   route = {
     title: 'Session Recordings',
@@ -543,9 +520,7 @@ export class FeatureRecordings implements TeleportFeature {
 }
 
 export class FeatureAudit implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Activity;
-  sideNavCategory = SideNavigationCategory.Audit;
+  category = NavigationCategory.Audit;
 
   route = {
     title: 'Audit Log',
@@ -570,9 +545,7 @@ export class FeatureAudit implements TeleportFeature {
 // - Clusters
 
 export class FeatureClusters implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Clusters;
-  sideNavCategory = SideNavigationCategory.Access;
+  category = NavigationCategory.Access;
 
   route = {
     title: 'Clusters',
@@ -601,9 +574,7 @@ export class FeatureClusters implements TeleportFeature {
 }
 
 export class FeatureTrust implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Clusters;
-  sideNavCategory = SideNavigationCategory.Access;
+  category = NavigationCategory.Access;
 
   route = {
     title: 'Trusted Root Clusters',
@@ -626,9 +597,7 @@ export class FeatureTrust implements TeleportFeature {
 }
 
 class FeatureDeviceTrust implements TeleportFeature {
-  category = NavigationCategory.Management;
-  section = ManagementSection.Identity;
-  sideNavCategory = SideNavigationCategory.Identity;
+  category = NavigationCategory.Identity;
   route = {
     title: 'Trusted Devices',
     path: cfg.routes.deviceTrust,
@@ -655,7 +624,7 @@ class FeatureDeviceTrust implements TeleportFeature {
 }
 
 class FeatureIntegrationStatus implements TeleportFeature {
-  category = NavigationCategory.Management;
+  category = NavigationCategory.Access;
 
   parent = FeatureIntegrations;
